Modify upstream implementation to relax ACL for view role
This patch fixes the [issue 1900][issue-1900] which requires `edit` role
to be removed from  `user`-jenkins namespace but it turns out that the
upstream [openshift-login-plugin][plugin] maps these roles to permissions
on jenkins and removing `edit` role leads to user losing permission to
start a build which prevents them from deploying to prod.

The patch addresses this issue by relaxing the ACL logic for `view` and
grants all capabilities that is granted for `edit` role to `view`.

See: [discussion on the issue][discussion]
Issue: [1900][issue-1900]

[issue-1900]: https://github.com/openshiftio/openshift.io/issues/1900
[discussion]: https://github.com/openshiftio/openshift.io/issues/1900#issuecomment-366928957
[plugin]: https://github.com/openshift/jenkins-openshift-login-plugin